# cloudhsm/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "KMS/SecretManager/CloudHSM API",
    "version": "1.2.0",
    "description": "さくらのクラウド KMS / SecretManager / CloudHSM API"
  },
  "servers": [
    {
      "url": "https://secure.sakura.ad.jp/cloud/zone/is1a/api/cloud/1.1",
      "description": "石狩第1ゾーン"
    },
    {
      "url": "https://secure.sakura.ad.jp/cloud/zone/is1b/api/cloud/1.1",
      "description": "石狩第2ゾーン"
    },
    {
      "url": "https://secure.sakura.ad.jp/cloud/zone/is1c/api/cloud/1.1",
      "description": "石狩第3ゾーン"
    },
    {
      "url": "https://secure.sakura.ad.jp/cloud/zone/tk1a/api/cloud/1.1",
      "description": "東京第1ゾーン"
    },
    {
      "url": "https://secure.sakura.ad.jp/cloud/zone/tk1b/api/cloud/1.1",
      "description": "東京第2ゾーン"
    }
  ],
  "paths": {
    "/cloudhsm/cloudhsms": {
      "get": {
        "operationId": "listCloudHSMs",
        "parameters": [
          {
            "name": "Count",
            "required": false,
            "in": "query",
            "description": "Number of results to return per page.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "From",
            "required": false,
            "in": "query",
            "description": "The initial index from which to return the results.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "tags": [
          "cloudhsm"
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCloudHSMList"
                }
              }
            },
            "description": ""
          }
        }
      },
      "post": {
        "operationId": "createCloudHSM",
        "tags": [
          "cloudhsm"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WrappedCreateCloudHSMRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedCreateCloudHSM"
                }
              }
            },
            "description": ""
          }
        }
      }
    },
...
    "/cloudhsm/cloudhsms/{resource_id}": {
      "get": {
        "operationId": "readCloudHSM",
        "parameters": [
          {
            "in": "path",
            "name": "resource_id",
            "schema": {
              "type": "string",
              "title": "リソースID"
            },
            "required": true
          }
        ],
        "tags": [
          "cloudhsm"
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedCloudHSM"
                }
              }
            },
            "description": ""
          }
        }
      },
      "put": {
        "operationId": "updateCloudHSM",
        "parameters": [
          {
            "in": "path",
            "name": "resource_id",
            "schema": {
              "type": "string",
              "title": "リソースID"
            },
            "required": true
          }
        ],
        "tags": [
          "cloudhsm"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WrappedCloudHSMRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedCloudHSM"
                }
              }
            },
            "description": ""
          }
        }
      },
      "delete": {
        "operationId": "deleteCloudHSM",
        "parameters": [
          {
            "in": "path",
            "name": "resource_id",
            "schema": {
              "type": "string",
              "title": "リソースID"
            },
            "required": true
          }
        ],
        "tags": [
          "cloudhsm"
        ],
        "responses": {
          "204": {
            "description": "No response body"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "basicAuth": []
    }
  ],
  "components": {
    "schemas": {
      "CloudHSM": {
        "type": "object",
        "properties": {
          "ID": {
            "type": "string",
            "readOnly": true,
            "example": "110000000000"
          },
          "CreatedAt": {
            "$ref": "#/components/schemas/DateTime"
          },
          "ModifiedAt": {
            "$ref": "#/components/schemas/DateTime"
          },
          "ServiceClass": {
            "enum": [
              "cloud/cloudhsm/partition"
            ],
            "type": "string",
            "description": "* `cloud/cloudhsm/partition` - Type-L7",
            "x-spec-enum-id": "aec33bb088a5c108",
            "readOnly": true,
            "title": "サービスクラス"
          },
          "Availability": {
            "enum": [
              "precreate",
              "available",
              "discontinued"
            ],
            "type": "string",
            "description": "* `precreate` - 準備中\n* `available` - 利用可能\n* `discontinued` - 廃止",
            "x-spec-enum-id": "329069c36229258c",
            "readOnly": true,
            "title": "状態"
          },
          "Name": {
            "type": "string",
            "title": "名前",
            "maxLength": 255
          },
          "Description": {
            "type": "string",
            "title": "説明"
          },
          "Tags": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "nullable": true
          },
          "IPv4NetworkAddress": {
            "type": "string",
            "title": "IPv4ネットワークアドレス",
            "example": "192.168.100.0"
          },
          "IPv4PrefixLength": {
            "type": "integer",
            "title": "IPv4プレフィックス長",
            "example": 24
          },
          "IPv4Address": {
            "type": "string",
            "readOnly": true,
            "example": "192.168.100.11"
          },
          "LocalRouter": {
            "allOf": [
              {
                "$ref": "#/components/schemas/LocalRouter"
              }
            ],
            "nullable": true,
            "readOnly": true
          },
          "InitialData": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CloudHSMInitialData"
              }
            ],
            "nullable": true,
            "readOnly": true
          }
        },
        "required": [
          "Availability",
          "CreatedAt",
          "Description",
          "ID",
          "InitialData",
          "IPv4Address",
          "IPv4NetworkAddress",
          "IPv4PrefixLength",
          "LocalRouter",
          "ModifiedAt",
          "Name",
          "ServiceClass",
          "Tags"
        ]
      },
...
      "CloudHSMInitialData": {
        "type": "object",
        "properties": {
          "PartitionName": {
            "type": "string",
            "title": "HSM 上のパーティション名"
          },
          "Certificate": {
            "type": "string",
            "title": "HSM のサーバー証明書",
            "example": "-----BEGIN CERTIFICATE-----\\nMIIC0DCCAbigAwIBAgIJAJ.....\\n-----END CERTIFICATE-----"
          },
          "PartitionID": {
            "type": "string",
            "title": "HSM 上のパーティション ID"
          }
        },
        "required": [
          "Certificate",
          "PartitionID",
          "PartitionName"
        ]
      },
...
      "CloudHSMRequest": {
        "type": "object",
        "properties": {
          "Name": {
            "type": "string",
            "minLength": 1,
            "title": "名前",
            "maxLength": 255
          },
          "Description": {
            "type": "string",
            "title": "説明"
          },
          "Tags": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "nullable": true
          },
          "IPv4NetworkAddress": {
            "type": "string",
            "minLength": 1,
            "title": "IPv4ネットワークアドレス",
            "example": "192.168.100.0"
          },
          "IPv4PrefixLength": {
            "type": "integer",
            "title": "IPv4プレフィックス長",
            "example": 24
          }
        },
        "required": [
          "IPv4NetworkAddress",
          "IPv4PrefixLength",
          "Name"
        ]
      },
...
      "CreateCloudHSM": {
        "type": "object",
        "properties": {
          "ID": {
            "type": "string",
            "readOnly": true,
            "example": "110000000000"
          },
          "CreatedAt": {
            "$ref": "#/components/schemas/DateTime"
          },
          "ModifiedAt": {
            "$ref": "#/components/schemas/DateTime"
          },
          "ServiceClass": {
            "enum": [
              "cloud/cloudhsm/partition"
            ],
            "type": "string",
            "description": "* `cloud/cloudhsm/partition` - Type-L7",
            "x-spec-enum-id": "aec33bb088a5c108",
            "readOnly": true,
            "title": "サービスクラス"
          },
          "Availability": {
            "enum": [
              "precreate",
              "available",
              "discontinued"
            ],
            "type": "string",
            "description": "* `precreate` - 準備中\n* `available` - 利用可能\n* `discontinued` - 廃止",
            "x-spec-enum-id": "329069c36229258c",
            "readOnly": true,
            "title": "状態"
          },
          "Name": {
            "type": "string",
            "title": "名前",
            "maxLength": 255
          },
          "Description": {
            "type": "string",
            "title": "説明"
          },
          "Tags": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "nullable": true
          },
          "IPv4NetworkAddress": {
            "type": "string",
            "title": "IPv4ネットワークアドレス",
            "example": "192.168.100.0"
          },
          "IPv4PrefixLength": {
            "type": "integer",
            "title": "IPv4プレフィックス長",
            "example": 24
          },
          "IPv4Address": {
            "type": "string",
            "readOnly": true,
            "example": "192.168.100.11"
          }
        },
        "required": [
          "IPv4NetworkAddress",
          "IPv4PrefixLength",
          "Name"
        ]
      },
...
      "CreateCloudHSMRequest": {
        "type": "object",
        "properties": {
          "Name": {
            "type": "string",
            "minLength": 1,
            "title": "名前",
            "maxLength": 255
          },
          "Description": {
            "type": "string",
            "title": "説明"
          },
          "Tags": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "nullable": true
          },
          "IPv4NetworkAddress": {
            "type": "string",
            "minLength": 1,
            "title": "IPv4ネットワークアドレス",
            "example": "192.168.100.0"
          },
          "IPv4PrefixLength": {
            "type": "integer",
            "title": "IPv4プレフィックス長",
            "example": 24
          }
        },
        "required": [
          "IPv4NetworkAddress",
          "IPv4PrefixLength",
          "Name"
        ]
      },
...
      "LocalRouter": {
        "type": "object",
        "properties": {
          "ResourceID": {
            "type": "string",
            "example": "110000000000"
          },
          "SecretKey": {
            "type": "string",
            "example": "PAIRINGSECRETKEY"
          }
        }
      },
...
      "PaginatedCloudHSMList": {
        "type": "object",
        "required": [
          "Count",
          "From",
          "Total",
          "CloudHSMs",
          "is_ok"
        ],
        "properties": {
          "Count": {
            "type": "integer",
            "example": 10
          },
          "From": {
            "type": "integer",
            "example": 0
          },
          "Total": {
            "type": "integer",
            "example": 10
          },
          "CloudHSMs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CloudHSM"
            }
          },
          "is_ok": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "WrappedCloudHSM": {
        "type": "object",
        "properties": {
          "CloudHSM": {
            "$ref": "#/components/schemas/CloudHSM"
          },
          "is_ok": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "CloudHSM",
          "is_ok"
        ]
      },
...
      "WrappedCloudHSMRequest": {
        "type": "object",
        "properties": {
          "CloudHSM": {
            "$ref": "#/components/schemas/CloudHSMRequest"
          }
        },
        "required": [
          "CloudHSM"
        ]
      },
...
      "WrappedCreateCloudHSM": {
        "type": "object",
        "properties": {
          "CloudHSM": {
            "$ref": "#/components/schemas/CreateCloudHSM"
          },
          "is_ok": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "CloudHSM",
          "is_ok"
        ]
      },
...
      "WrappedCreateCloudHSMRequest": {
        "type": "object",
        "properties": {
          "CloudHSM": {
            "$ref": "#/components/schemas/CreateCloudHSMRequest"
          }
        },
        "required": [
          "CloudHSM"
        ]
      },
...
      "DateTime": {
        "type": "string",
        "format": "iso8601",
        "example": "2025-02-05T12:19:22.551827+09:00"
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "APIキーとシークレットをご利用ください"
      }
    }
  }
}